day13.
A lot of LINQ usage. Feels like solutions are easier to think through a functional lense.

Approach was through working with a list of points, then to a 2d array because I wanted to know quickly the bounds of the page and then back to list of points because I settled on computaing pageSize in the beginning and kept it updated through folds.

Tricket part was getting the new pageSize correct. `pageSize.x - f.fold - 1` was key.