A tough one.
Obviously I cannot do a string expansion.

The idea here was to expand in one direction, get the results and use that to continuing tallying the results.
To make it faster, I skip levels.
2 iterations
NNCB -> NB + CCNBBBCBHCB
NB -> NBB
NB -> NBB
..

utitlize caching to increase performance as well. If we already know a string combo at a certain level expands with a given frequency map, reuse that result.

I am caculating and recursing in a DFS way.